refactor: ChatRoom에서 MessageList 로직 분리
